Rename SplatDate to SplatCalendar
